Fix: Set strictNullChecks true in tsconfig.json
When strictNullChecks is false, null and undefined are effectively ignored by the language. This can lead to unexpected errors at runtime.

When strictNullChecks is true, null and undefined have their own distinct types and you’ll get a type error if you try to use them where a concrete value is expected.

- reference
  - referencehttps://www.typescriptlang.org/tsconfig#strictNullChecks